EmsSinkException has been moved from abstract class to trait for pattern match notifications.
EmsSinkTask now creates an execution context for the upload.

Tests were added for:
  schemaless JSON
  JSON payloads.

For schemaless JSON the connector infers the schema on a message by messge base. It's how the data is deserialized and we cannot identify if a field is nullable or not.

Cleans up the sbt project to avoid packing dependencies which are not used.