Support for obfuscating fields
With this PR the connector allows new configurations:

* connect.ems.obfuscation.fields
* connect.ems.obfuscation.sha512.salt
* connect.ems.obfuscation.method

If the fields is specified it will take the obfuscation method.And if obfuscation is SHA512 then it will require the salt.

With the obfuscation fields enabled it will apply the data content change *only* for strings. If a path is not found it won't fail the conversion - this is due to JSON inputs where a schema is determined at the record level.