## Description
There is a requirement to "humanize" the obfuscation output. This allows the users of EMS to link the obfuscated value in EMS with the real value in their system. The code change is limited to DataObfuscation. The obfuscation output is represented as hex string.

SBT version has been updated to 1.6.2

A few files were not formatted or had the header, hence the changes for those.
## Testing
The unit tests have been updated to reflect the latest changes.