Schema compatibility e2e test (#32)
* Fix changing schema

When a schema change happens and there's a drop in a column from the previous one the sink would fail. The exception generated by the Parquet library is a consequence of schema fields and value alignment.

The code change is targeting the WriterManager. When a schema rollover is required, it should not use the current writer state since it carries the old schema.