Orderable Field (Rebase of Stefan's original PR) (#28)
* Orderable field

To support deduplication on Event Collection the data uploaded requires a field which is sortable. This will allow to pick for a given key the latest value.

The way this is injected is to use the Kafka topic-partition-offset implicit order. At the partition level the messages are ordered. The message timestamp is not reliable, because it can be for the same record messages to have the values reverted or equal.

The way it's build is to start with a large number which is multiplied by the partition and then add the offset. the offset in kafka starts at 0 and increases. It's a long. This ensures for the same partition, all records are orderd. And it's expected the data in Kafka is ordered for a given primary key.

Introduces a new connector config to allow the user to overwrite the auto inserted field when PK(s) is/are used. Whenever a PK(s) list is used we need a field which is sortable such that many records with the same PK in a file uploaded can be deduped on EMS. The connector inserts the Kafka message offset as the long sortable field.