[CKC-58]Primary Keys null validation (#21)
#### Description

Currently, the connector allows the user to pick up the primary keys list. However, a primary key value cannnot and should not be null.

This PR introduces an extra validation in PrimaryKeysValidator class to ensure a PK field value is not null. In case it is, an error will be returned.

#### Tests

Unit tests coverage was added to confirm the null-pk checks.

Co-authored-by: stheppi <jenkins@lenses.io>
Co-authored-by: Andrea Fiore <andrea.giulio.fiore@gmail.com>